Backend

+++ app/src/main/java/ru/netology/dao/PostDao.kt
@@ -16,10 +16,10 @@ interface PostDao {
     fun insert(post: PostEntity)
 
     @Query("UPDATE PostEntity SET content = :content WHERE id = :id")
-    fun updateContentById(id: Int, content: String)
+    fun updateContentById(id: Long, content: String)
 
     fun save(post: PostEntity) =
-        if (post.id == 0) insert(post) else updateContentById(post.id, post.content)
+        if (post.id == 0L) insert(post) else updateContentById(post.id, post.content)
 
     @Query("""
         UPDATE PostEntity SET
@@ -27,7 +27,7 @@ interface PostDao {
         likedByMe = CASE WHEN likedByMe THEN 0 ELSE 1 END
         WHERE id = :id
         """)
-    fun likeById(id: Int)
+    fun likeById(id: Long)
 
     @Query("""
        UPDATE PostEntity SET
@@ -35,8 +35,8 @@ interface PostDao {
         WHERE id = :id
     """
     )
-    fun shareById(id: Int)
+    fun shareById(id: Long)
 
     @Query("DELETE FROM PostEntity WHERE id = :id")
-    fun removeById(id: Int)
+    fun removeById(id: Long)
 }
